requirements.txt
accelerate==1.2.0
addict==2.4.0
av==14.0.1
bitsandbytes==0.45.0
clip_anytorch==2.6.0
datasets
diffusers==0.27.2
einops==0.8.0
fsspec==2024.12.0
ftfy==6.3.1
gradio==3.28.3
huggingface_hub==0.26.5
imageio==2.36.1
matplotlib
nltk==3.9.1
numpy
omegaconf==2.3.0
onnxruntime==1.19.2
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
packaging==24.2
pandas==2.2.3
peft==0.14.0
Pillow==11.1.0
PyYAML==6.0.2
regex==2024.11.6
Requests==2.32.3
safetensors==0.5.2
seaborn==0.13.2
segment_anything==1.0
supervision==0.25.1
termcolor==2.5.0
torchmetrics==1.6.1
tqdm==4.67.1
tracker==0.1.1
transformers==4.45.2
typer==0.15.1

urllib3>=1.26,<3
idna>=3.4
charset_normalizer>=3.3.0
certifi>=2023.7.22
chardet>=5.2.0
psutil>=5.9
pynvml>=11.5.0
pytz

tokenizers==0.20
python-dateutil
six
